pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "pyspark-dubber"
dynamic = ["version"]
requires-python = ">= 3.10"
authors = [
    { name = "Francesco Pasa", email = "francescopasa@gmail.com" },
    { name = "Marco Lotz" },
]
description = """
A compatibility layer for the pyspark API, allowing you to run pyspark code on
backends such as DuckDB and Polars without porting your code.
"""
readme = "README.md"
license = "LGPL-3.0-or-later"

dependencies = [
    "ibis-framework[duckdb]",
    "pandas",
    "pyarrow",
    "pyarrow-hotfix",
    "sqlglot[rs]",
    "python-dateutil",
    "tzlocal",
    "lark",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pyspark",
    "setuptools",  # pandas still depends on it post 3.12 because it sucks
    "mkdocs-material",
]

[project.urls]
documentation = "https://frapa.github.io/pyspark-dubber/"

[tool.hatch.version]
path = "pyspark_dubber/__init__.py"